+++ NEWS.md
@@ -1,5 +1,13 @@
 <!-- NEWS.md is maintained by https://cynkra.github.io/fledge, do not edit -->
 
+# tiltIndicator 0.0.0.9069
+
+* `pstr_prepare_scenario()` is now named `xstr_prepare_scenario()` (#385).
+
+# tiltIndicator 0.0.0.9068
+
+* `pstr_polish_output_at_copmany_level()` is now named `xstr_polish_output_at_copmany_level()` (#383).
+
 # tiltIndicator 0.0.0.9067
 
 * `xstr_prune_companies()` now expects the column `company_id` (#380).
@@ -47,7 +55,7 @@ an error (#311).
 
 # tiltIndicator 0.0.0.9057
 
-* `pstr_prepare_scenarios()` now handles "weo" data correctly (@kalashsinghal #309).
+* `pstr_prepare_scenario()` now handles "weo" data correctly (@kalashsinghal #309).
 
 # tiltIndicator 0.0.0.9056